Flask Server/requirements.txt
accelerate==0.29.3
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
beautifulsoup4==4.10.0
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.502
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.5.14
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fastapi==0.110.1
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.48
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.45
Lazify==0.4.0
literalai==0.0.500
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.6.3
numpy==1.26.4
odfpy==1.4.1
openai==1.25.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pdfminer.six==20200517
pdfminer3k==1.3.4
pdfplumber==0.5.28
pillow==10.3.0
ply==3.11
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.0b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==2.10.5
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
seaborn==0.13.2
sentencepiece==0.2.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
syncer==2.0.3
tabulate==0.8.9
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
Wand==0.6.13
watchdog==4.0.0
watchfiles==0.20.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1